backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
openai==1.70.0
python-dotenv==1.0.1
python-multipart
